refactor(functions): extract route-planner helpers into shared module
Move metresBetween, midpoint, filterToEllipse, beamSearch (→ beamSearchOrienteering),
and finalise (→ finaliseRoute) from scripts/plan_route.ts into the new pure
helper module _routePlanner.ts. Add filterToCorridor (corridor / cross-track
filter needed by the upcoming routePostboxes Cloud Function). Smoke test
output unchanged (36 points, 5 stops). Six new mocha tests for filterToCorridor
all passing (274 total).
